@@ -1,6 +1,6 @@
 class Category < ApplicationRecord
 	belongs_to :user
-	belongs_to :group
+	belongs_to :group, optional: true
 	has_many :events
 	has_and_belongs_to_many :repeat_exceptions
 
