@@ -5,7 +5,7 @@ class Event < ApplicationRecord
   belongs_to :user
   alias_attribute :creator, :user
 
-  belongs_to :group
+  belongs_to :group, optional: true
   belongs_to :category
   has_and_belongs_to_many :repeat_exceptions
 
