@@ -64,7 +64,7 @@ class User < ApplicationRecord
   ##########################
 
   # Returns a url to the avatar with the width in pixels.
-	def avatar_url(size)
+	def avatar_url(size = 256)
     return "#{image_url.split("?")[0]}?sz=#{size}" if has_google_avatar? # google avatar
     return avatar.url(size <= 60 ? :thumb : :profile) if avatar.exists? # uploaded avatar
     gravatar_url(size) # default avatar
@@ -220,51 +220,4 @@ class User < ApplicationRecord
   ##########################
   ## END GEN USER METHODS ##
   ##########################
-
-  ##########################
-  ##     MISC METHODS     ##
-  ##########################
-
-  # Convert the user into a hash with the least data needed to show search
-  # Recall that clients can see the JSON with a bit of inspection, so only
-  # public information should be included here
-  def convert_to_json
-    user_obj = {} #create a hash representing the user
-
-    # Required fields for search/tokenInput - name and image url
-    user_obj[:name] = self.name
-    user_obj[:image_url] = self.avatar_url(50)
-    user_obj[:model_name] = "User" # specify what type of object this is (used for site search, which handles many object types)
-
-    user_obj #and return the user
-  end
-
-  # Ranks a collection of users for a given search query
-  def self.rank(users, query)
-    query = query.downcase
-
-    users.sort {|a,b| b.rank(query) <=> a.rank(query) } # return users by rank descending (hence sort is flipped)
-  end
-
-  # Ranks the user based on the query
-  # The ranking prioritizes the first name starting with the query
-  # then the middle/last name starting with the query
-  # and finally the query being included at all (handled by SQL)
-  def rank(query)
-    score = 0 # default score is 0
-
-    if name.downcase.starts_with?(query)
-      score = 2
-    elsif name.downcase.include?(" " + query)
-      score = 1
-    end
-
-    return score
-  end
-
-  ##########################
-  ##   END MISC METHODS   ##
-  ##########################
-
-
 end
