@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180703195151) do
+ActiveRecord::Schema.define(version: 2018_09_14_143025) do
 
   create_table "categories", force: :cascade do |t|
     t.string "name"
