@@ -1,3 +1,30 @@
+# Teaspoon doesn't allow you to pass client driver options to the Selenium WebDriver. This monkey patch
+# is a temporary fix until this PR is merged: https://github.com/jejacks0n/teaspoon/pull/519.
+require 'teaspoon/driver/selenium'
+
+Teaspoon::Driver::Selenium.class_eval do
+  def run_specs(runner, url)
+    driver =
+      if driver_options[:options]
+        ::Selenium::WebDriver.for(driver_options[:client_driver], options: driver_options[:options])
+      else
+        ::Selenium::WebDriver.for(driver_options[:client_driver])
+      end
+
+    driver.navigate.to(url)
+
+    ::Selenium::WebDriver::Wait.new(driver_options).until do
+      done = driver.execute_script("return window.Teaspoon && window.Teaspoon.finished")
+      driver.execute_script("return window.Teaspoon && window.Teaspoon.getMessages() || []").each do |line|
+        runner.process("#{line}\n")
+      end
+      done
+    end
+  ensure
+    driver.quit if driver
+  end
+end
+
 Teaspoon.configure do |config|
   # Determines where the Teaspoon routes will be mounted. Changing this to "/jasmine" would allow you to browse to
   # `http://localhost:3000/jasmine` to run your tests.
@@ -65,7 +92,7 @@ Teaspoon.configure do |config|
     #suite.hook :fixtures, &proc{}
 
     # Determine whether specs loaded into the test harness should be embedded as individual script tags or concatenated
-    # into a single file. Similar to Rails' asset `debug: true` and `config.assets.debug = true` options. By default, 
+    # into a single file. Similar to Rails' asset `debug: true` and `config.assets.debug = true` options. By default,
     # Teaspoon expands all assets to provide more valuable stack traces that reference individual source files.
     #suite.expand_assets = true
 
@@ -98,7 +125,8 @@ Teaspoon.configure do |config|
   # Selenium Webdriver: https://github.com/modeset/teaspoon/wiki/Using-Selenium-WebDriver
   # BrowserStack Webdriver: https://github.com/modeset/teaspoon/wiki/Using-BrowserStack-WebDriver
   # Capybara Webkit: https://github.com/modeset/teaspoon/wiki/Using-Capybara-Webkit
-  #config.driver = :phantomjs
+
+  config.driver = :selenium
 
   # Specify additional options for the driver.
   #
@@ -106,7 +134,13 @@ Teaspoon.configure do |config|
   # Selenium Webdriver: https://github.com/modeset/teaspoon/wiki/Using-Selenium-WebDriver
   # BrowserStack Webdriver: https://github.com/modeset/teaspoon/wiki/Using-BrowserStack-WebDriver
   # Capybara Webkit: https://github.com/modeset/teaspoon/wiki/Using-Capybara-Webkit
-  #config.driver_options = nil
+
+  config.driver_options = {
+    client_driver: :chrome,
+    options: Selenium::WebDriver::Chrome::Options.new(
+      args: ['headless', 'no-sandbox', 'disable-dev-shm-usage', 'remote-debugging-port=9222']
+    )
+  }
 
   # Specify the timeout for the driver. Specs are expected to complete within this time frame or the run will be
   # considered a failure. This is to avoid issues that can arise where tests stall.
